Fix warning.
git-svn-id: file:///home/yoann/SVN/libprelude/trunk@11095 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/prelude-plugin.c
@@ -845,7 +845,7 @@ void prelude_plugin_entry_set_plugin(prelude_plugin_entry_t *pe, prelude_plugin_
 
 void prelude_plugin_set_preloaded_symbols(void *symlist)
 {
-        size_t len;
+        unsigned long len;
         lt_dlsymlist *s = symlist;
         static lt_dlsymlist rpl_sym[65535] = {
                 { "@PROGNAME@", NULL },
